# api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: hello@permify.co
    name: API Support
    url: https://github.com/Permify/permify/issues
  description: Permify is an open source authorization service for creating fine-grained
    and scalable authorization systems.
  license:
    name: Apache-2.0 license
    url: https://github.com/Permify/permify/blob/master/LICENSE
  title: Permify API
  version: v0.9.9
servers:
- url: /
...
paths:
...
  /v1/tenants/{tenant_id}/schemas/list:
    post:
      operationId: schemas.list
      parameters:
      - description: "Identifier of the tenant, if you are not using multi-tenancy\
          \ (have only one tenant) use pre-inserted tenant <code>t1</code> for this\
          \ field. Required, and must match the pattern \\“[a-zA-Z0-9-,]+\\“, max\
          \ 64 bytes."
        in: path
        name: tenant_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/schemas_list_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaListResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: An unexpected error response.
      summary: list schema
      tags:
      - Schema
      x-codeSamples:
      - label: go
        lang: go
        source: |-
          sr, err: = client.Schema.List(context.Background(), &v1.SchemaListRequest {
              TenantId: "t1",
              PageSize: "10",
              ContinuousToken: "",
          })
      - label: node
        lang: javascript
        source: |-
          let res = client.schema.list({
              tenantId: "t1",
              continuousToken: ""
          })
      - label: cURL
        lang: curl
        source: |-
          curl --location --request POST 'localhost:3476/v1/tenants/{tenant_id}/schemas/read' \
          --header 'Content-Type: application/json' \
          --data-raw '{
              "page_size": "10",
              "continuous_token": ""
          }'
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts:
      - application/json
  /v1/tenants/{tenant_id}/schemas/partial-write:
    patch:
      operationId: schemas.partial-write
      parameters:
      - description: |-
          tenant_id is a string that identifies the tenant. It must match the pattern "[a-zA-Z0-9-,]+",
          be a maximum of 64 bytes, and must not be empty.
        in: path
        name: tenant_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/It_contains_the_tenant_id_to_identify_the_tenant_and_metadata_of_the_schema_to_be_edited__with_the_corresponding_edits_to_various_entities'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaPartialWriteResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: An unexpected error response.
      summary: partially update your authorization model
      tags:
      - Schema
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts:
      - application/json
  /v1/tenants/{tenant_id}/schemas/read:
    post:
      operationId: schemas.read
      parameters:
      - description: "Identifier of the tenant, if you are not using multi-tenancy\
          \ (have only one tenant) use pre-inserted tenant <code>t1</code> for this\
          \ field. Required, and must match the pattern \\“[a-zA-Z0-9-,]+\\“, max\
          \ 64 bytes."
        in: path
        name: tenant_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/schemas_read_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaReadResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: An unexpected error response.
      summary: read schema
      tags:
      - Schema
      x-codeSamples:
      - label: go
        lang: go
        source: |-
          sr, err: = client.Schema.Read(context.Background(), &v1.SchemaReadRequest {
              TenantId: "t1",
              Metadata: &v1.SchemaReadRequestMetadata{
                  SchemaVersion: "cnbe6se5fmal18gpc66g",
              },
          })
      - label: node
        lang: javascript
        source: |-
          let res = client.schema.read({
              tenantId: "t1",
              metadata: {
              schemaVersion: swResponse.schemaVersion,
              },
              })
      - label: cURL
        lang: curl
        source: |-
          curl --location --request POST 'localhost:3476/v1/tenants/{tenant_id}/schemas/read' \
          --header 'Content-Type: application/json' \
          --data-raw '{
              "metadata": {
                  "schema_version": "cnbe6se5fmal18gpc66g"
              }
          }'
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts:
      - application/json
  /v1/tenants/{tenant_id}/schemas/write:
    post:
      operationId: schemas.write
      parameters:
      - description: "Identifier of the tenant, if you are not using multi-tenancy\
          \ (have only one tenant) use pre-inserted tenant <code>t1</code> for this\
          \ field. Required, and must match the pattern \\“[a-zA-Z0-9-,]+\\“, max\
          \ 64 bytes."
        in: path
        name: tenant_id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/schemas_write_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaWriteResponse'
          description: A successful response.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
          description: An unexpected error response.
      summary: write schema
      tags:
      - Schema
      x-codeSamples:
      - label: go
        lang: go
        source: |-
          sr, err: = client.Schema.Write(context.Background(), &v1.SchemaWriteRequest {
              TenantId: "t1",
              Schema: `
              "entity user {}\n\n    entity organization {\n\n        relation admin @user\n        relation member @user\n\n        action create_repository = (admin or member)\n        action delete = admin\n    }\n\n    entity repository {\n\n        relation owner @user\n        relation parent @organization\n\n        action push = owner\n        action read = (owner and (parent.admin and parent.member))\n        action delete = (parent.member and (parent.admin or owner))\n    }"
              `,
          })
      - label: node
        lang: javascript
        source: |-
          client.schema.write({
              tenantId: "t1",
              schema: `
              "entity user {}\n\n    entity organization {\n\n        relation admin @user\n        relation member @user\n\n        action create_repository = (admin or member)\n        action delete = admin\n    }\n\n    entity repository {\n\n        relation owner @user\n        relation parent @organization\n\n        action push = owner\n        action read = (owner and (parent.admin and parent.member))\n        action delete = (parent.member and (parent.admin or owner))\n    }"
              `
          }).then((response) => {
              // handle response
          })
      - label: cURL
        lang: curl
        source: |-
          curl --location --request POST 'localhost:3476/v1/tenants/{tenant_id}/schemas/write' \
          --header 'Content-Type: application/json' \
          --data-raw '{
              "schema": "entity user {}\n\n    entity organization {\n\n        relation admin @user\n        relation member @user\n\n        action create_repository = (admin or member)\n        action delete = admin\n    }\n\n    entity repository {\n\n        relation owner @user\n        relation parent @organization\n\n        action push = owner\n        action read = (owner and (parent.admin and parent.member))\n        action delete = (parent.member and (parent.admin or owner))\n }"
          }'
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
  schemas:
...
    SchemaListResponse:
      example:
        head: head
        continuous_token: continuous_token
        schemas:
        - created_at: created_at
          version: version
        - created_at: created_at
          version: version
      properties:
        head:
          title: head of the schemas is the latest version available for the tenant
          type: string
        schemas:
          items:
            $ref: '#/components/schemas/SchemaList'
          title: list of schema versions with creation timestamps
          type: array
        continuous_token:
          description: continuous_token is a string that can be used to paginate and
            retrieve the next set of results.
          type: string
      title: |-
        SchemaListResponse is the response message for the List method in the Schema service.
        It returns a paginated list of schemas
      type: object
...
    Status:
      example:
        code: 0
        details:
        - '@type': '@type'
        - '@type': '@type'
        message: message
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
        details:
          items:
            $ref: '#/components/schemas/Any'
          type: array
      type: object
...
    schemas_list_request:
      description: |-
        SchemaListRequest is the request message for the List method in the Schema service.
        It contains tenant_id for which the schemas are to be listed.
      properties:
        page_size:
          description: |-
            page_size is the number of tenants to be returned in the response.
            The value should be between 1 and 100.
          format: int64
          type: integer
        continuous_token:
          description: |-
            continuous_token is an optional parameter used for pagination.
            It should be the value received in the previous response.
          type: string
      type: object
    ? It_contains_the_tenant_id_to_identify_the_tenant_and_metadata_of_the_schema_to_be_edited__with_the_corresponding_edits_to_various_entities
    : properties:
        metadata:
          $ref: '#/components/schemas/SchemaPartialWriteRequestMetadata'
        partials:
          additionalProperties:
            $ref: '#/components/schemas/Partials'
          title: Map of entity name with the values needed to be updated
          type: object
      title: |-
        It contains the tenant_id to identify the tenant and metadata of the schema to be edited,
        with the corresponding edits to various entities
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: Authorization
      type: apiKey
...